LIX-390 # Replace Object.assign with object spread and split the Oxlint plugin
