+++ services/typescript-quality-runner/src/stylelint/plugin.ts
