+++ services/typescript-quality-runner/test-quality.sh
@@ -9,10 +9,10 @@ tool_dir="/usr/src/quality-runner"
 runner_dir="$tool_dir"
 dprint_bin="$tool_dir/node_modules/.bin/dprint"
 oxlint_bin="$tool_dir/node_modules/.bin/oxlint"
-import_order_checker="$runner_dir/import-specifier-order.ts"
-source_extension_runner="$runner_dir/source-extension-runner.ts"
-typescript_format_runner="$runner_dir/typescript-format-runner.ts"
-stylelint_runner="$tool_dir/stylelint-runner.ts"
+import_order_checker="$runner_dir/src/formatting/import-specifier-order.ts"
+source_extension_runner="$runner_dir/src/source-extensions/source-extension-runner.ts"
+typescript_format_runner="$runner_dir/src/formatting/typescript-format-runner.ts"
+stylelint_runner="$tool_dir/src/stylelint/runner.ts"
 # dprint formats stylesheets only. TypeScript goes through the Oxfmt-backed formatter, so a
 # TypeScript path passed to dprint intentionally matches no plugin.
 dprint_config="$tool_dir/dprint.json"
